# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: RESTful API for managing todos in a MERN stack application
  version: 1.0.0
  contact:
    name: Nadeesha Medagama
    email: nadeeshamedagama@example.com
servers:
  - url: https://todo-backend.choreo.dev/api
    description: Production server
  - url: http://localhost:5000/api
    description: Development server
paths:
  /todos:
    get:
      summary: Get all todos
      description: Retrieve a list of all todos
      tags:
        - Todos
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Todo'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new todo
      description: Create a new todo item
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Todo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Todo:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the todo
        text:
          type: string
          description: The todo text content
        completed:
          type: boolean
          description: Whether the todo is completed
        createdAt:
          type: string
          format: date-time
          description: When the todo was created
        updatedAt:
          type: string
          format: date-time
          description: When the todo was last updated
      required:
        - text
        - completed
    CreateTodoRequest:
      type: object
      properties:
        text:
          type: string
          description: The todo text content
          minLength: 1
          maxLength: 500
        completed:
          type: boolean
          description: Whether the todo is completed
          default: false
      required:
        - text
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
      required:
        - success
        - error
...